backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
apscheduler>=3.10.0
anthropic>=0.40.0
openai>=1.0.0
supabase>=2.0.0
elevenlabs>=1.0.0
feedparser>=6.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
